Backend

+++ backend/util/file_management.js
@@ -53,12 +53,20 @@ const generateJSON = (name, data) => {
     fs.writeFileSync(name, JSON.stringify(data, null, 4));
 }
 
+// Generate a file of any format
+const generateFile = (name, data) => {
+    fs.writeFileSync(name, data);
+}
+
 // Read a csv file
-const readCSV = (path) => new Promise((resolve, reject) => {
-    // If the file path does not end in '.csv', delete file and throw error
+const readCSV = (path, del = true) => new Promise((resolve, reject) => {
     if (path.substring(path.length - 4, path.length) !== ".csv") {
+        // If the file path does not end in '.csv', delete file and throw error
         fs.unlinkSync(path);
         throw new Error(`${ path.substring(path.length - 4, path.length) } is an invalid file type`);
+    } else if (!fileExists(path)) {
+        // If the file does not exist, throw error
+        throw new Error(`${ path } does not exist`);
     }
 
     // Parse csv
@@ -67,11 +75,17 @@ const readCSV = (path) => new Promise((resolve, reject) => {
         .pipe(csv_read())
         .on("data", d => data.push(d))
         .on("end", () => {
-            fs.unlinkSync(path);
+            if (del) {
+                fs.unlinkSync(path);
+            }
+
             resolve(data);
         })
         .on("error", err => {
-            fs.unlinkSync(path);
+            if (del) {
+                fs.unlinkSync(path);
+            }
+
             reject(err);
         });
 });
@@ -104,6 +118,7 @@ module.exports = {
     fileExists,
     generateCSV,
     generateJSON,
+    generateFile,
     readCSV,
     readJSON,
     uploadFile
